// src/app/api/models/poll-results-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PollOptionResultDto } from '../models/poll-option-result-dto';

/**
 * DTO for poll results
 */
export interface PollResultsDto {

  /**
   * Whether multiple answers are allowed
   */
  allowMultipleAnswers?: boolean;

  /**
   * Date and time when the poll was created
   */
  createdAt?: string;

  /**
   * User ID who created the poll
   */
  createdBy?: string;

  /**
   * Message ID containing the poll
   */
  messageId?: string;

  /**
   * Poll options with vote counts
   */
  options?: Array<PollOptionResultDto>;

  /**
   * Poll question
   */
  question?: string;

  /**
   * Total number of unique voters
   */
  totalVoters?: number;

  /**
   * Total number of votes cast
   */
  totalVotes?: number;
}